fuck around and find out!

+++ src/main/java/de/bixilon/minosoft/protocol/packets/s2c/play/map/MapS2CP.kt
@@ -1,6 +1,6 @@
 /*
  * Minosoft
- * Copyright (C) 2020-2025 Moritz Zwerger
+ * Copyright (C) 2020-2023 Moritz Zwerger
  *
  * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
  *
@@ -13,7 +13,7 @@
 
 package de.bixilon.minosoft.protocol.packets.s2c.play.map
 
-import de.bixilon.kmath.vec.vec2.i.Vec2i
+import de.bixilon.kotlinglm.vec2.Vec2i
 import de.bixilon.minosoft.data.world.map.MapPin
 import de.bixilon.minosoft.data.world.map.MapPinTypes
 import de.bixilon.minosoft.protocol.packets.s2c.PlayS2CPacket
@@ -44,7 +44,7 @@ class MapS2CP(buffer: PlayInByteBuffer) : PlayS2CPacket {
         for (i in 0 until pinCount) {
             if (buffer.versionId < V_18W19A) {
                 val rawDirection = buffer.readUnsignedByte()
-                val position = Vec2i(buffer.readByte().toInt(), buffer.readByte().toInt())
+                val position = Vec2i(buffer.readByte(), buffer.readByte())
 
                 val direction: Int
                 val type: MapPinTypes
@@ -60,7 +60,7 @@ class MapS2CP(buffer: PlayInByteBuffer) : PlayS2CPacket {
                 continue
             }
             val type = MapPinTypes[buffer.readVarInt()]
-            val position = Vec2i(buffer.readByte().toInt(), buffer.readByte().toInt())
+            val position = Vec2i(buffer.readByte(), buffer.readByte())
             val direction = buffer.readUnsignedByte()
             val displayName = buffer.readOptional { buffer.readChatComponent() }
             pins[position] = MapPin(direction, type, displayName)
